predictor/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
psycopg2-binary
scikit-learn
numpy
